PropagationUtil split between BinaryPropagationHandler (for the interface definition) and BinaryPropagationHandlerImpl (for the implementation).
