+++ benchmarks/build.gradle
@@ -23,7 +23,8 @@ jmh {
 }
 
 dependencies {
-    compile project(':instrumentation-java-core')
+    compile project(':instrumentation-java-core'),
+            project(':instrumentation-java-core-impl')
 }
 
 compileJmhJava {
