+++ core/src/main/java/com/google/instrumentation/trace/BinaryPropagationHandler.java
@@ -18,44 +18,47 @@ import static com.google.common.base.Preconditions.checkNotNull;
 import java.text.ParseException;
 
 /**
- * This is a helper class for {@link SpanContext} propagation on the wire.
+ * This is a helper class for {@link SpanContext} propagation on the wire using binary encoding.
  *
- * <p>Binary format:
+ * <p>Example of usage on the client:
  *
- * <ul>
- * <li>Binary value: &lt;version_id&gt;&lt;version_format&gt;
- * <li>version_id: 1-byte representing the version id.
- * <li>For version_id = 0:
- *     <ul>
- *     <li>version_format: &lt;field&gt;&lt;field&gt;
- *     <li>field_format: &lt;field_id&gt;&lt;field_format&gt;
- *     <li>Fields:
- *         <ul>
- *         <li>TraceId: (field_id = 0, len = 16, default = &#34;0000000000000000&#34;) - 16-byte
- *             array representing the trace_id.
- *         <li>SpanId: (field_id = 1, len = 8, default = &#34;00000000&#34;) - 8-byte array
- *             representing the span_id.
- *         <li>TraceOptions: (field_id = 2, len = 1, default = &#34;0&#34;) - 1-byte array
- *             representing the trace_options.
- *         </ul>
- *     <li>Fields MUST be encoded using the field id order (smaller to higher).
- *     <li>Valid value example:
- *         <ul>
- *         <li>{0, 0, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 1, 97, 98, 99,
- *             100, 101, 102, 103, 104, 2, 1}
- *         <li>version_id = 0;
- *         <li>trace_id = {64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79}
- *         <li>span_id = {97, 98, 99, 100, 101, 102, 103, 104};
- *         <li>trace_options = {1};
- *         </ul>
- *     </ul>
+ * <pre>{@code
+ * private static final Tracer tracer = TraceService.getTracer();
+ * private static final BinaryPropagationHandler binaryPropagationHandler =
+ *     TraceService.getBinaryPropagationHandler();
+ * void onSendRequest() {
+ *   try (NonThrowingCloseable ss = tracer.spanBuilder("Sent.MyRequest").startScopedSpan()) {
+ *     byte[] binaryValue = binaryPropagationHandler.toBinaryValue(
+ *         tracer.getCurrentContext().context());
+ *     // Send the request including the binaryValue and wait for the response.
+ *   }
+ * }
+ * }</pre>
  *
- * </ul>
+ * <p>Example of usage on the server:
+ *
+ * <pre>{@code
+ * private static final Tracer tracer = TraceService.getTracer();
+ * private static final BinaryPropagationHandler binaryPropagationHandler =
+ *     TraceService.getBinaryPropagationHandler();
+ * void onRequestReceived() {
+ *   // Get the binaryValue from the request.
+ *   SpanContext spanContext = SpanContext.INVALID;
+ *   try {
+ *     if (binaryValue != null) {
+ *       binaryPropagationHandler.fromBinaryValue(binaryValue);
+ *     }
+ *   } catch (ParseException e) {
+ *     // Maybe log the exception.
+ *   }
+ *   try (NonThrowingCloseable ss =
+ *            tracer.spanBuilderWithRemoteParent(spanContext, "Recv.MyRequest").startScopedSpan()) {
+ *     // Handle request and send response back.
+ *   }
+ * }
+ * }</pre>
  */
-public final class BinaryPropagationHandler {
-  // Mask to extract a byte value.
-  private static volatile BinaryHandler binaryHandler = DefaultBinaryHandler.INSTANCE;
-
+public abstract class BinaryPropagationHandler {
   /**
    * Serializes a {@link SpanContext} using the binary format.
    *
@@ -63,10 +66,7 @@ public final class BinaryPropagationHandler {
    * @return the serialized binary value.
    * @throws NullPointerException if the {@code spanContext} is {@code null}.
    */
-  public static byte[] toBinaryValue(SpanContext spanContext) {
-    checkNotNull(spanContext, "spanContext");
-    return binaryHandler.toBinaryFormat(spanContext);
-  }
+  public abstract byte[] toBinaryValue(SpanContext spanContext);
 
   /**
    * Parses the {@link SpanContext} from the binary format.
@@ -76,106 +76,27 @@ public final class BinaryPropagationHandler {
    * @throws NullPointerException if the {@code input} is {@code null}.
    * @throws ParseException if the version is not supported or the input is invalid
    */
-  public static SpanContext fromBinaryValue(byte[] bytes) throws ParseException {
-    checkNotNull(bytes, "bytes");
-    return binaryHandler.fromBinaryFormat(bytes);
+  public abstract SpanContext fromBinaryValue(byte[] bytes) throws ParseException;
+
+  static BinaryPropagationHandler getNoopBinaryPropagationHandler() {
+    return NoopBinaryPropagationHandler.INSTANCE;
   }
 
-  /**
-   * Sets a new {@link BinaryHandler}.
-   *
-   * @param newHandler the new {@code Handler} to be set.
-   */
-  static void setBinaryHandler(BinaryHandler newHandler) {
-    binaryHandler = newHandler;
-  }
-
-  /** Abstract class that allows implementation of the new version format support. */
-  public abstract static class BinaryHandler {
-    /**
-     * Serializes a {@link SpanContext} using the binary format.
-     *
-     * @param spanContext the {@code SpanContext} to serialize.
-     * @return the serialized binary value.
-     */
-    public abstract byte[] toBinaryFormat(SpanContext spanContext);
-
-    /**
-     * Parses the {@link SpanContext} from the binary format.
-     *
-     * @param bytes a binary encoded buffer from which the {@code SpanContext} will be parsed.
-     * @return the parsed {@code SpanContext}.
-     * @throws ParseException is any parsing error.
-     */
-    public abstract SpanContext fromBinaryFormat(byte[] bytes) throws ParseException;
-  }
-
-  /** Version 0 implementation of the {@code VersionHandler}. */
-  public static final class DefaultBinaryHandler extends BinaryHandler {
-    /** Singleton instance of this class. */
-    public static final DefaultBinaryHandler INSTANCE = new DefaultBinaryHandler();
-
-    private static final byte VERSION_ID = 0;
-    private static final int VERSION_ID_OFFSET = 0;
-    // The version_id/field_id size in bytes.
-    private static final byte ID_SIZE = 1;
-    private static final byte TRACE_ID_FIELD_ID = 0;
-    private static final int TRACE_ID_FIELD_ID_OFFSET = VERSION_ID_OFFSET + ID_SIZE;
-    private static final int TRACE_ID_OFFSET = TRACE_ID_FIELD_ID_OFFSET + ID_SIZE;
-    private static final byte SPAN_ID_FIELD_ID = 1;
-    private static final int SPAN_ID_FIELD_ID_OFFSET = TRACE_ID_OFFSET + TraceId.SIZE;
-    private static final int SPAN_ID_OFFSET = SPAN_ID_FIELD_ID_OFFSET + ID_SIZE;
-    private static final byte TRACE_OPTION_FIELD_ID = 2;
-    private static final int TRACE_OPTION_FIELD_ID_OFFSET = SPAN_ID_OFFSET + SpanId.SIZE;
-    private static final int TRACE_OPTIONS_OFFSET = TRACE_OPTION_FIELD_ID_OFFSET + ID_SIZE;
-    private static final int FORMAT_LENGTH =
-        4 * ID_SIZE + TraceId.SIZE + SpanId.SIZE + TraceOptions.SIZE;
+  private static final class NoopBinaryPropagationHandler extends BinaryPropagationHandler {
+    static final NoopBinaryPropagationHandler INSTANCE = new NoopBinaryPropagationHandler();
 
     @Override
-    public byte[] toBinaryFormat(SpanContext spanContext) {
-      byte[] bytes = new byte[FORMAT_LENGTH];
-      bytes[VERSION_ID_OFFSET] = VERSION_ID;
-      bytes[TRACE_ID_FIELD_ID_OFFSET] = TRACE_ID_FIELD_ID;
-      spanContext.getTraceId().copyBytesTo(bytes, TRACE_ID_OFFSET);
-      bytes[SPAN_ID_FIELD_ID_OFFSET] = SPAN_ID_FIELD_ID;
-      spanContext.getSpanId().copyBytesTo(bytes, SPAN_ID_OFFSET);
-      bytes[TRACE_OPTION_FIELD_ID_OFFSET] = TRACE_OPTION_FIELD_ID;
-      spanContext.getTraceOptions().copyBytesTo(bytes, TRACE_OPTIONS_OFFSET);
-      return bytes;
+    public byte[] toBinaryValue(SpanContext spanContext) {
+      checkNotNull(spanContext, "spanContext");
+      return new byte[0];
     }
 
     @Override
-    public SpanContext fromBinaryFormat(byte[] bytes) throws ParseException {
-      if (bytes.length == 0 || bytes[0] != VERSION_ID) {
-        throw new ParseException("Unsupported version.", 0);
-      }
-      TraceId traceId = TraceId.INVALID;
-      SpanId spanId = SpanId.INVALID;
-      TraceOptions traceOptions = TraceOptions.DEFAULT;
-      int pos = 1;
-      try {
-        if (bytes.length > pos && bytes[pos] == TRACE_ID_FIELD_ID) {
-          traceId = TraceId.fromBytes(bytes, pos + ID_SIZE);
-          pos += ID_SIZE + TraceId.SIZE;
-        }
-        if (bytes.length > pos && bytes[pos] == SPAN_ID_FIELD_ID) {
-          spanId = SpanId.fromBytes(bytes, pos + ID_SIZE);
-          pos += ID_SIZE + SpanId.SIZE;
-        }
-        if (bytes.length > pos && bytes[pos] == TRACE_OPTION_FIELD_ID) {
-          traceOptions = TraceOptions.fromBytes(bytes, pos + ID_SIZE);
-        }
-        return new SpanContext(traceId, spanId, traceOptions);
-      } catch (ArrayIndexOutOfBoundsException e) {
-        throw new ParseException("Invalid input: " + e.toString(), pos);
-      } catch (IndexOutOfBoundsException e) {
-        throw new ParseException("Invalid input: " + e.toString(), pos);
-      }
+    public SpanContext fromBinaryValue(byte[] bytes) throws ParseException {
+      checkNotNull(bytes, "bytes");
+      return SpanContext.INVALID;
     }
 
-    private DefaultBinaryHandler() {}
+    private NoopBinaryPropagationHandler() {}
   }
-
-  // Disallow instances of this class.
-  private BinaryPropagationHandler() {}
 }
